Changed City so the costs it saves are doubles instead of integers

+++ aima-core/src/main/java/aima/core/environment/tsp/ArcsAverageHeuristicFunction.java
@@ -20,7 +20,7 @@ public class ArcsAverageHeuristicFunction extends AbstractTSPHeuristicFunction {
 			
 			for(City to : toVisit) {
 				
-				Integer cost = from.getCost(to);
+				Double cost = from.getCost(to);
 				
 				if(cost != null) {
 					sumArcs += cost;
